inc: minimal changes to let ICS build
This is about the bare minimum change needed to get inc running on ICS.
It builds and boots, but almost nothing works.  It's a start.

Most hardware support will need BOARD_USES_QCOM_HARDWARE, but that makes
the build crash and burn, so I've left it out for now.

IDC files were nabbed from Evervolv; with BOARD_USE_LEGACY_TOUCHSCREEN
this lets the touchscreen work on the existing kernel.

Invalid overlay values have been commented out for now, until I
determine which are obsolete or just renamed to something else.

Change-Id: I367c447705ea3fd1034245d909237b67e6986d6c